+++ src/barretenberg/waffle/stdlib/uint/uint.hpp
@@ -67,13 +67,13 @@ template <typename Composer, typename Native> class uint {
     uint operator|(const uint& other) const;
     uint operator~() const;
 
-    uint operator>>(const uint64_t shift) const;
-    uint operator<<(const uint64_t shift) const;
+    uint operator>>(const size_t shift) const;
+    uint operator<<(const size_t shift) const;
 
-    uint ror(const uint64_t target_rotation) const;
-    uint rol(const uint64_t target_rotation) const;
-    uint ror(const uint256_t target_rotation) const { return ror(target_rotation.data[0]); }
-    uint rol(const uint256_t target_rotation) const { return rol(target_rotation.data[0]); }
+    uint ror(const size_t target_rotation) const;
+    uint rol(const size_t target_rotation) const;
+    uint ror(const uint256_t target_rotation) const { return ror(static_cast<size_t>(target_rotation.data[0])); }
+    uint rol(const uint256_t target_rotation) const { return rol(static_cast<size_t>(target_rotation.data[0])); }
 
     bool_t<Composer> operator>(const uint& other) const;
     bool_t<Composer> operator<(const uint& other) const;
@@ -93,8 +93,8 @@ template <typename Composer, typename Native> class uint {
     uint operator^=(const uint& other) { *this = operator^(other); return *this; }
     uint operator|=(const uint& other) { *this = operator|(other); return *this; }
 
-    uint operator>>=(const uint64_t shift) { *this = operator>>(shift); return *this; }
-    uint operator<<=(const uint64_t shift) { *this = operator<<(shift); return *this; }
+    uint operator>>=(const size_t shift) { *this = operator>>(shift); return *this; }
+    uint operator<<=(const size_t shift) { *this = operator<<(shift); return *this; }
 
     uint normalize() const;
 
@@ -132,7 +132,7 @@ private:
         AND,
         XOR,
     };
-    
+
     std::pair<uint, uint> divmod(const uint& other) const;
     uint logic_operator(const uint& other, const LogicOp op_type) const;
     uint weak_normalize() const;
